Move slot scheduler under state
