+++ services/beaconchain/src/main/java/tech/pegasys/artemis/services/beaconchain/BeaconChainService.java
@@ -24,6 +24,7 @@ import tech.pegasys.artemis.Constants;
 import tech.pegasys.artemis.pow.api.ChainStartEvent;
 import tech.pegasys.artemis.pow.api.ValidatorRegistrationEvent;
 import tech.pegasys.artemis.services.ServiceInterface;
+import tech.pegasys.artemis.state.SlotScheduler;
 import tech.pegasys.artemis.state.StateTreeManager;
 
 public class BeaconChainService implements ServiceInterface {
